[Refactor][Core/Dashboard] Make datacenter.py a private implementation for NodeHead
